Implemented CRUD for Courses and navigation between courses and students

+++ maven-javafx-infc20-assignment/src/main/java/se/ics/lu/models/Course.java
@@ -35,7 +35,7 @@ public class Course {
         return credits;
     }
 
-    public void setCredits(int credits) {
+    public void setCredits(double credits) {
         this.credits = credits;
     }
